[fix] contrast on selection text
As pointed out by user Soiled-Linen (aka. Lint), the website's current selection background being yellow makes selected text difficult to view. Lint suggested fixing this by changing the text colour to a dark red.

This commit instead keeps the text colour as white, and sets the _background_ colour to a somewhat lighter dark red, which makes highlighting consistent with the colour pattern for link hovers and tab selection on the website (both of which are white text on a red background). The new background colour `#A0303B` has been tested on multiple websites to pass AA- and AAA-compliance for normal and large text.