[ci] Move defaults to top-level
Despite the GH documentation suggesting that this was the correct way to
do things for one job, the CI runner complained; might be something to
do with Pelican's action, which we're using? In any case, a top-level
defaults declaration should definitely be okay, and there is really no
reason not to: the other directories in the repo are irrelevant wrt.
building the site.