[release] Update the Download page for 0.7.0
With the previous commit uploading the tarball, checksum, and the actual
blog post announcing the release, this commit updates the Downloads page
to point people at it. Relatedly, it also moves 0.6.0 to the "previous
releases" part of the page.

I am undecided as to whether this is sufficient to remove the disclaimer
at the top of the webpage. For now, I will leave it in place, just to be
safe.

There was also some housekeeping done in terms of:
* The Chez Scheme URL being moved since the project went open-source.
* Minor typos and phrasings to do with the legacy version (Idris 1).