[ index ] Update community references
The `idris-hackers` org does not appear to be active outside of hosting
`idris-mode` for emacs (which, at the time of writing, seems to be
different from `idris2-mode` in that it supports both Idris1 and
Idris2).

Some current community efforts were under `idris-community`, but `pack`
also enabled having a broad variety of repo hosts, hence why I've
referenced the pack-db repo.