[ news ] Announce release of Idris2 v0.8.0
Idris2 version 0.8.0 has been released almost 2 years after the previous
release.  This puts the tarball on the website, alongside a little
announcement for folk who follow the atom feed or who check the website
for news.

The tarball and its SHA256-checksum are different from the ones found on
GitHub, because the `mkdist.sh` script removes some unnecessary files
and directories before bundling.